requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic[email]==2.5.0
streamlit==1.28.2
plotly==5.18.0
pandas==2.1.3
python-multipart==0.0.6
feedparser==6.0.10
beautifulsoup4==4.12.2
openai==1.6.1
